обновлена сборка образа: процесс унифицирован

+++ scripts/clear.sh
@@ -2,10 +2,30 @@
 
 yarn run clean
 
-rm ./license.txt
-rm ./lerna-debug.log
-rm ./docker-error.log
-rm ./package-json.lock
-rm -r ./node_modules
-rm -r ./.node_modules
-rm -rf ./.nx
+if [[ -f "./license.txt" ]]; then
+    rm ./license.txt
+fi
+
+if [[ -f "./lerna-debug.log" ]]; then
+    rm ./lerna-debug.log
+fi
+
+if [[ -f "./docker-error.log" ]]; then
+    rm ./docker-error.log
+fi
+
+if [[ -f "./package-json.lock" ]]; then
+    rm ./package-json.lock
+fi
+
+if [[ -d "./node_modules" ]]; then
+    rm -r ./node_modules
+fi
+
+if [[ -d "./.node_modules" ]]; then
+    rm -r ./.node_modules
+fi
+
+if [[ -d "./.nx" ]]; then
+    rm -rf ./.nx
+fi
